Proper shiny stats tracking?
I'm taking a stab at computing shiny stats more accurately.

This PR computes shiny stats, but does so by the combination
of 'encounterId,accountUsername'. We need to look at every
unique encounterId/accountUsername pair, so this is hooked
in before we check if an encounter is a duplicate.

For prometheus, this sends shundo,nundo stats by area, like the
original commented out metrics. But for 'any shiny', I'm sending
metrics by area and dex Id. And non-shiny counts are submitted
as well by area and dex Id. The purpose for this is so that we
can compute accurate shiny odds. I did not do this for the DB
stats.

This also makes shutdown cleanly shutdown the stats writers before
the program exists.